Tony added lots of thoughts to README.md.
Tony tends to abuse README.mds, and use them as note pads to share ideas
with collaborators. He wrote ideas about an outline of a GA algorithm,
what inputs we coud use, what outputs we could use, the Knight's desire
to keep the grail away from the center. The Grail's desire to stay
toward the center, and division of labor.

Signed-off-by: TexasTony <tknapp@conncoll.edu>