Updated README, Changed wall to (- 20000).
I deleted most of the old notes from the readme, updated the input
section, and made a breif description of our efforts of the
evening/night/early morning.

In agent, in (get-block-vlaue), I changed the return for an out of
bounds block to (- 20000), that way it is a negative value.