Change GA so that 10 grids are generated per generation
The sum of the iterations of these runs are then subtracted from 5010
(501 * 10). This is the fitness for the DNA strand. This will prevent
the DNA from having an amazing fitness if the Grail and the Knight start
next to each other. 

I am currently using random-seed to control the formation of the grids,
but that may be interefearing with the gene mutations.